Separate read and process data functions.
Divides read data and process data functions into two classes.  One that
reads in the data and a second one that process the data for peak
detection.
